requirements.txt
bcrypt==3.2.0
bidict==0.21.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-SocketIO==5.1.0
gevent==21.1.2
greenlet==1.1.0
gunicorn==20.1.0
h11==0.12.0
harperdb==0.4.2
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycparser==2.20
PyJWT==2.1.0
python-dotenv==0.17.1
python-engineio==4.2.0
python-socketio==5.3.0
requests==2.25.1
simple-websocket==0.2.0
six==1.16.0
urllib3==1.26.5
Werkzeug==2.0.1
wsproto==1.0.0
zope.event==4.5.0
zope.interface==5.4.0
